+++ packages/lumx-vue/src/components/text-field/RawInputText.tsx
@@ -16,6 +16,10 @@ export const emitSchema = {
     change: (value: string, _name?: string, _event?: Event) => typeof value === 'string',
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     input: (value: string, _name?: string, _event?: Event) => typeof value === 'string',
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    focus: (_event?: FocusEvent) => true,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    blur: (_event?: FocusEvent) => true,
 };
 
 /**
@@ -40,6 +44,14 @@ const RawInputText = defineComponent(
             emit('input', value, name, event);
         };
 
+        const handleFocus = (event?: any) => {
+            emit('focus', event);
+        };
+
+        const handleBlur = (event?: any) => {
+            emit('blur', event);
+        };
+
         return () => {
             return (
                 <RawInputTextUI
@@ -51,6 +63,8 @@ const RawInputText = defineComponent(
                     theme={props.theme || defaultTheme.value}
                     handleChange={handleChange}
                     handleInput={handleInput}
+                    handleFocus={handleFocus}
+                    handleBlur={handleBlur}
                 />
             );
         };